requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
flask_apscheduler
git+https://github.com/communitiesuk/funding-service-design-utils.git@tferns-patch-1-move-sqs-helpers-to-utils
alembic==1.11.1
    # via flask-migrate
attrs==21.4.0
    # via jsonschema
babel==2.10.3
    # via flask-babel
beautifulsoup4==4.12.2
    # via
    #   -r requirements.in
    #   funding-service-design-utils
blinker==1.6.2
    # via sentry-sdk
boto3==1.26.140
    # via flipper-client
botocore==1.29.140
    # via
    #   boto3
    #   s3transfer
cachetools==5.3.0
    # via flipper-client
certifi==2022.6.15
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
chardet==4.0.0
    # via prance
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   clickclick
    #   flask
clickclick==20.10.2
    # via connexion
commonmark==0.9.1
    # via rich
connexion==2.14.2
    # via -r requirements.in
cryptography==37.0.4
    # via pyjwt
docopt==0.6.2
    # via notifications-python-client
flask==2.2.5
    # via
    #   -r requirements.in
    #   connexion
    #   flask-babel
    #   flask-migrate
    #   flask-redis
    #   flask-sqlalchemy
    #   funding-service-design-utils
    #   sentry-sdk
flask-babel==2.0.0
    # via funding-service-design-utils
flask-migrate==4.0.4
    # via funding-service-design-utils
flask-redis==0.4.0
    # via funding-service-design-utils
flask-sqlalchemy==3.0.3
    # via
    #   flask-migrate
    #   funding-service-design-utils
flask-talisman==0.8.1
    # via -r requirements.in
flipper-client==1.3.2
    # via funding-service-design-utils
greenlet==3.0.0
    # via sqlalchemy
gunicorn==20.1.0
    # via funding-service-design-utils
idna==3.3
    # via requests
inflection==0.5.1
    # via connexion
itsdangerous==2.1.2
    # via
    #   connexion
    #   flask
jinja2==3.1.2
    # via
    #   flask
    #   flask-babel
    #   swagger-ui-bundle
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.6.1
    # via
    #   connexion
    #   openapi-schema-validator
    #   openapi-spec-validator
mako==1.2.4
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
    #   sentry-sdk
    #   werkzeug
notifications-python-client==6.3.0
    # via -r requirements.in
openapi-schema-validator==0.2.3
    # via openapi-spec-validator
openapi-spec-validator==0.4.0
    # via -r requirements.in
packaging==21.3
    # via connexion
prance==0.21.8.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyee==6.0.0
    # via flipper-client
pygments==2.12.0
    # via rich
pyjwt[crypto]==2.4.0
    # via
    #   funding-service-design-utils
    #   notifications-python-client
pyparsing==3.0.9
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-consul==1.1.0
    # via flipper-client
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.20.0
    # via funding-service-design-utils
python-json-logger==2.0.2
    # via funding-service-design-utils
pytz==2022.1
    # via
    #   babel
    #   flask-babel
    #   funding-service-design-utils
pyyaml==6.0
    # via
    #   clickclick
    #   connexion
    #   funding-service-design-utils
    #   openapi-spec-validator
redis==4.5.5
    # via
    #   flask-redis
    #   flipper-client
requests==2.28.1
    # via
    #   -r requirements.in
    #   connexion
    #   funding-service-design-utils
    #   notifications-python-client
    #   prance
    #   python-consul
rich==12.4.4
    # via funding-service-design-utils
ruamel-yaml==0.17.21
    # via prance
s3transfer==0.6.1
    # via boto3
semver==2.13.0
    # via prance
sentry-sdk[flask]==1.31.0
    # via funding-service-design-utils
six==1.16.0
    # via
    #   flask-talisman
    #   prance
    #   python-consul
    #   python-dateutil
    #   thrift
soupsieve==2.4.1
    # via beautifulsoup4
sqlalchemy==2.0.15
    # via
    #   alembic
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via funding-service-design-utils
swagger-ui-bundle==0.0.9
    # via -r requirements.in
thrift==0.16.0
    # via flipper-client
typing-extensions==4.6.2
    # via
    #   alembic
    #   sqlalchemy
urllib3==1.26.12
    # via
    #   botocore
    #   requests
    #   sentry-sdk
werkzeug==2.2.3
    # via
    #   connexion
    #   flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
